+++ src/App.css
@@ -6,14 +6,13 @@ body,
 }
 
 h1, h2, h3 {
-  font-family: 'Oswald', 'Arial Narrow', sans-serif;
+  font-family: 'Open Sans', sans-serif;
 }
 
 code {
   display: block;
   margin: 20px 0 15px 0;
   padding: 10px;
-  background: #eee;
 }
 
 .container {
@@ -22,35 +21,33 @@ code {
   padding: 45px 20px;
 }
 
-.pure-button-primary {
-  background-color: #0c1a2b;
-}
-
-.pure-button-primary:hover {
-  background-color: #233e5e;
-}
-
-.pure-form input[type="text"]:focus {
-  border-color: #0c1a2b;
-}
-
 /* NAVBAR */
 
 .navbar {
+  background: none;
+  display: flex;
   position: fixed;
-  padding: 5px;
-  background: #0c1a2b;
-  font-family: 'Oswald', 'Arial Narrow', sans-serif;
+  padding: 20px 0;
+  font-family: 'Open Sans', sans-serif;
 }
 
-.navbar a {
-  color: #fff;
+.nav-link {
+  color: rgb(57, 57, 57);
+  font-size: 18px;
+  justify-content: space-between;
+  padding: 20px 20px;
+  text-decoration: none;
 }
+
+.space-between {
+  width: 80%;
+}
+
 
 .navbar a:active,
 .navbar a:focus,
 .navbar a:hover {
-  background: #233e5e;
+  color: #1b1b1b;
 }
 
 .navbar .pure-menu-heading {
@@ -76,4 +73,56 @@ code {
 
 .tile {
   width: 300px;
+}
+
+html {
+  background: #fff;
+}
+
+body {
+  font-family: 'Open-sans', sans-serif;
+  font-weight: lighter;
+  text-align: center;
+  color: #000;
+}
+
+.page-title {
+  font-family: 'Open-sans', sans-serif;
+  font-size: 50px;
+  font-weight: 300;
+  margin: 20px;
+  padding-top: 60px;
+}
+
+.input-field {
+  margin: 10px 0;
+  padding: 10px 20px;
+  border: 1px solid rgb(193, 193, 193);
+  border-radius: 5px;
+  outline: 0;
+  width: 300px;
+  font-size: 16px;
+}
+
+.btn-submit {
+  background-color: #fff;
+  border-radius: 5px;
+  color: rgb(57, 57, 57);
+  font-size: 18px;
+  font-weight: 300;
+  margin: 5px;
+  padding: 8px;
+}
+
+.btn-submit:hover {
+  background-color: rgb(211, 227, 255);
+}
+
+.header-slide.horizTranslate {
+  -webkit-transition: 3s;
+  -moz-transition: 3s;
+  -ms-transition: 3s;
+  -o-transition: 3s;
+  transition: 3s;
+  margin-bottom: 50% !important;
 }
